refactor(workspaces): split cmux/tmux adapters into separate files; drop cmux SSH inheritance
Make the Workspace module navigable by moving its two backends into their
own files behind the existing in-core adapter interface, and remove the
narrow cmux SSH-remote inheritance path.

- Extract the shared Adapter interface, public types, and the
  runWorkspaceCommand/isSignalAborted helpers into workspaceAdapter.ts.
- Move the cmux backend to cmuxAdapter.ts and the tmux backend to
  tmuxAdapter.ts. workspaces.ts becomes a thin facade: it re-exports the
  public types, resolves which backend to use (resolveWorkspaceKind),
  caches it per config, and exposes the workspaces API. External import
  sites are unchanged.
- Delete the cmux SSH-remote inheritance path: probeCurrentCmuxRemote
  (the `cmux current-workspace` probe), buildSshWrappedCommand (the
  ssh-wrapped launch-command builder), and the CMUX_WORKSPACE_ID gating.
  cmux open() now always passes --cwd/--command directly.

No behavior change for open/list/close/probe/accessHint on either backend
beyond the removed SSH path. cmux status-pill painting is retained; tmux
Workspaces still live in the dedicated `groundcrew` tmux session. This is
cleanup, not an extension point — no plugin contract is introduced.

Closes staff-1038
